feat: Allows independent control of DB driver +more
In mongo deployments, it's useful to access the underlying transaction.
In non-mongo scenarios, it's useful to swap the entire DB layer out for
another engine capable of the same operations such as Postgres. This
set of changes makes it easy to change the DB Driver.

Additional features include the "next" schedule being passed in the
ack, fail, and dead events. Index management is now local to the DB
driver.

BREAKING: This results in a change to how Queues are instantiated. Mongo
is now also listed as an optional peer dependency.