chore: Moves to tsup
This is an update to the build system and artifacts. No import paths
were changed. This switches out unbuild (rollup based) for tsup (esbuild
based) for bundling the CJS/ESM artifacts. This allows us to colocate
the tsup config in package.json next to our export manifest, making it
easier to ensure all exports are generated. Syncpack was updated to
make the order easier to follow (scripts>tsup>exports>files).